pyproject.toml
[project]
name = "sheep-classification-challenge"
description = "A computer vision project to classify Arabian sheep breeds from images using machine learning. Built for a cultural challenge celebrating Eid al-Adha"
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.10"

dependencies = [
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.52.4",
    "openai-clip>=1.0.1",
    "pytorch-lightning>=2.5.1.post0",
    "pandas>=2.3.0",
    "matplotlib>=3.10.3",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",

]

[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "ipython",
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-coverage",
    "types-PyYAML",
    "kaggle", # Kaggle API
    "ipykernel" # Jupyter Notebooks
]

[tool.pydoclint]
style = "sphinx"
exclude = "test_"
allow-init-docstring = true
arg-type-hints-in-docstring = false
check-return-types = false
require-return-section-when-returning-nothing = false

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
ignore_missing_imports = true

[tool.setuptools]
packages = ["src"]

[tool.setuptools_scm]
# can be empty if no extra settings are needed, presence enables setuptools_scm

[tool.setuptools.dynamic]
version = {attr = "src.version.__version__"}
